+++ R/utils_labels.R
@@ -3,12 +3,29 @@
 # to the transformed vector
 
 #' @keywords internal
-.set_back_labels <- function(new, old, include_values = TRUE) {
+.set_back_labels <- function(new, old, include_values = TRUE, reverse_values = FALSE) {
   # labelled data?
   attr(new, "label") <- attr(old, "label", exact = TRUE)
-  labels <- attr(old, "labels", exact = TRUE)
-  if (isTRUE(include_values) && !is.null(labels)) {
-    attr(new, "labels") <- stats::setNames(rev(labels), names(labels))
+  value_labels <- attr(old, "labels", exact = TRUE)
+  # "include_values" is used to preserve value labels
+  if (isTRUE(include_values) && !is.null(value_labels)) {
+    if (reverse_values) {
+      # reverse values? Used for "reverse_scale()"
+      attr(new, "labels") <- stats::setNames(rev(value_labels), names(value_labels))
+    } else {
+      # keep value oder? Used for "to_numeric()"
+      if (is.numeric(new)) {
+        if (any(grepl("[^0-9]", value_labels))) {
+          # if we have any non-numeric characters, convert to numeric
+          attr(new, "labels") <- stats::setNames(as.numeric(as.factor(value_labels)), names(value_labels))
+        } else {
+          # if we have numeric, or "numeric character" (like "1", "2", "3" etc.)
+          attr(new, "labels") <- stats::setNames(as.numeric(value_labels), names(value_labels))
+        }
+      } else {
+        attr(new, "labels") <- stats::setNames(value_labels, names(value_labels))
+      }
+    }
   } else if (isFALSE(include_values)) {
     attr(new, "labels") <- NULL
   }
@@ -30,7 +47,7 @@
   # check positions of matching values and levels
   levels_in_labs <- stats::na.omit(match(value_labels, levels(x)))
   labs_in_levels <- stats::na.omit(match(levels(x), value_labels))
-  # sanity check - if labelled values and levels don't match
+  # validation check - if labelled values and levels don't match
   if (!length(levels_in_labs) || !length(labs_in_levels)) {
     if (verbose) {
       insight::format_alert(