+++ tests/testthat/test-data_reverse.R
@@ -335,16 +335,20 @@ test_that("reverse_scale warns if single value to reverse", {
 test_that("reverse_scale select helpers", {
   data(iris)
   out <- rescale(iris, to = list(
-    "Sepal.Length" = c(0, 1),
-    "Petal.Length" = c(-1, 0)
+    Sepal.Length = c(0, 1),
+    Petal.Length = c(-1, 0)
   ), select = ends_with("length"))
 
   expect_identical(out$Sepal.Length, iris$Sepal.Length, tolerance = 1e-3)
 
-  out <- rescale(iris, to = list(
-    "Sepal.Length" = c(0, 1),
-    "Petal.Length" = c(-1, 0)
-  ), select = ends_with("length"), ignore_case = TRUE)
+  out <- rescale(iris,
+    to = list(
+      Sepal.Length = c(0, 1),
+      Petal.Length = c(-1, 0)
+    ),
+    select = ends_with("length"),
+    ignore_case = TRUE
+  )
 
   expect_identical(head(out$Sepal.Length), c(0.22222, 0.16667, 0.11111, 0.08333, 0.19444, 0.30556), tolerance = 1e-3)
 })